sensors: Fix flakiness in the same-origin access check tests
Commit f73dfaffe ("[sensors] Check that sensor frame and focused frame of
same origin") added a few same-origin access check tests for sensors in
generic-sensor-iframe-tests.sub.js.

The test that checks that a sensor is not suspended when focus changes to
another same-origin frame was flaky though, especially on the Mac bots.

The actual fix is a one-liner: we were missing a call to Sensor.stop() so
that the sensor in the top-level document wouldn't continue to push
'reading' events to our EventWatcher while we were waiting for the
"is_sensor_suspended" command to be processed in iframe_sensor_handler.html.

I'm also taking the opportunity to make a few other minor improvements:
* Fix a few typos in the comments, and improve the comment explaining why we
  can end up sending more "reading" events than originally expected.
* Create the sensor belonging to the top-level document with a higher
  frequency than the default 5Hz. This makes it a lot easier to reproduce
  the error we were having and will hopefully make the test consistently
  fail in the future instead of being flaky should another related bug
  occur.
* Make the EventWatcher instances we create also watch for "error" events;
  if an error event ever gets sent, the test should now fail instead of just
  timing out waiting for a "reading" event that is never delivered.

Bug: 855222, 857520
Change-Id: I945bb19d74e608089aa9aee82a4e74e05fae37ae